+++ libspu/core/value.h
@@ -38,11 +38,13 @@ struct ValueProto {
 
 class Value final {
   NdArrayRef data_;
+  std::optional<NdArrayRef> imag_;
   DataType dtype_ = DT_INVALID;
 
  public:
   Value() = default;
   explicit Value(NdArrayRef data, DataType dtype);
+  explicit Value(NdArrayRef real, NdArrayRef imag, DataType dtype);
 
   /// Forward ndarray methods.
   inline int64_t numel() const { return data_.numel(); }
@@ -58,6 +60,9 @@ class Value final {
   const NdArrayRef& data() const { return data_; }
   NdArrayRef& data() { return data_; }
 
+  const std::optional<NdArrayRef>& imag() const { return imag_; }
+  std::optional<NdArrayRef>& imag() { return imag_; }
+
   // Get vtype, is readonly and decided by the underline secure compute engine.
   Visibility vtype() const;
   bool isPublic() const { return vtype() == VIS_PUBLIC; }
@@ -67,6 +72,7 @@ class Value final {
   DataType dtype() const { return dtype_; }
   bool isInt() const { return isInteger(dtype()); }
   bool isFxp() const { return isFixedPoint(dtype()); }
+  bool isComplex() const { return imag_.has_value(); }
 
   // Set dtype.
   //