+++ libspu/mpc/cheetah/nonlinear/compare_prot_test.cc
@@ -18,9 +18,8 @@
 
 #include "gtest/gtest.h"
 
-#include "libspu/core/xt_helper.h"
 #include "libspu/mpc/cheetah/ot/basic_ot_prot.h"
-#include "libspu/mpc/semi2k/type.h"
+#include "libspu/mpc/cheetah/type.h"
 #include "libspu/mpc/utils/ring_ops.h"
 #include "libspu/mpc/utils/simulate.h"
 
@@ -43,44 +42,111 @@ INSTANTIATE_TEST_SUITE_P(
 
 TEST_P(CompareProtTest, Basic) {
   size_t kWorldSize = 2;
-  int64_t n = 16;
+  Shape shape = {13, 2, 3};
   FieldType field = std::get<0>(GetParam());
   size_t radix = std::get<2>(GetParam());
   bool greater_than = std::get<1>(GetParam());
 
   NdArrayRef inp[2];
-  inp[0] = ring_rand(field, {n});
-  inp[1] = ring_rand(field, {n});
+  inp[0] = ring_rand(field, shape);
+  inp[1] = ring_rand(field, shape);
 
   DISPATCH_ALL_FIELDS(field, "", [&]() {
-    auto xinp = xt_mutable_adapt<ring2k_t>(inp[0]);
+    auto xinp = NdArrayView<ring2k_t>(inp[0]);
     xinp[0] = 1;
     xinp[1] = 10;
     xinp[2] = 100;
 
-    xinp = xt_mutable_adapt<ring2k_t>(inp[1]);
+    xinp = NdArrayView<ring2k_t>(inp[1]);
     xinp[0] = 1;
     xinp[1] = 9;
     xinp[2] = 1000;
   });
 
-  ArrayRef cmp_oup[2];
+  NdArrayRef cmp_oup[2];
   utils::simulate(kWorldSize, [&](std::shared_ptr<yacl::link::Context> ctx) {
     auto conn = std::make_shared<Communicator>(ctx);
     int rank = ctx->Rank();
     auto base = std::make_shared<BasicOTProtocols>(conn);
     CompareProtocol comp_prot(base, radix);
-    auto _c = comp_prot.Compute(flatten(inp[rank]), greater_than);
+    auto _c = comp_prot.Compute(inp[rank], greater_than);
     cmp_oup[rank] = _c;
   });
 
   DISPATCH_ALL_FIELDS(field, "", [&]() {
-    auto xout0 = ArrayView<ring2k_t>(cmp_oup[0]);
-    auto xout1 = ArrayView<ring2k_t>(cmp_oup[1]);
-    auto xinp0 = xt_adapt<ring2k_t>(inp[0]);
-    auto xinp1 = xt_adapt<ring2k_t>(inp[1]);
+    auto xout0 = NdArrayView<ring2k_t>(cmp_oup[0]);
+    auto xout1 = NdArrayView<ring2k_t>(cmp_oup[1]);
+    auto xinp0 = NdArrayView<ring2k_t>(inp[0]);
+    auto xinp1 = NdArrayView<ring2k_t>(inp[1]);
 
-    for (int64_t i = 0; i < n; ++i) {
+    for (int64_t i = 0; i < shape.numel(); ++i) {
+      bool expected = greater_than ? xinp0[i] > xinp1[i] : xinp0[i] < xinp1[i];
+      bool got_cmp = xout0[i] ^ xout1[i];
+      ASSERT_EQ(expected, got_cmp);
+    }
+  });
+}
+
+TEST_P(CompareProtTest, SpecifiedBitwidth) {
+  size_t kWorldSize = 2;
+  FieldType field = std::get<0>(GetParam());
+  size_t radix = std::get<2>(GetParam());
+  bool greater_than = std::get<1>(GetParam());
+  int64_t bw = 16;
+
+  NdArrayRef inp[2];
+  int64_t n = 1 << 18;
+  inp[0] = ring_rand(field, {n * 2});
+  inp[1] = ring_rand(field, {n * 2});
+
+  DISPATCH_ALL_FIELDS(field, "", [&]() {
+    ring2k_t mask = (static_cast<ring2k_t>(1) << bw) - 1;
+    auto xinp = NdArrayView<ring2k_t>(inp[0]);
+    xinp[0] = 1;
+    xinp[1] = 10;
+    xinp[2] = 100;
+    pforeach(0, inp[0].numel(), [&](int64_t i) { xinp[i] &= mask; });
+
+    xinp = NdArrayView<ring2k_t>(inp[1]);
+    xinp[0] = 1;
+    xinp[1] = 9;
+    xinp[2] = 1000;
+    pforeach(0, inp[0].numel(), [&](int64_t i) { xinp[i] &= mask; });
+  });
+
+  inp[0] = inp[0].reshape({n, 2});
+  inp[1] = inp[1].reshape({n, 2});
+
+  NdArrayRef cmp_oup[2];
+  utils::simulate(kWorldSize, [&](std::shared_ptr<yacl::link::Context> ctx) {
+    auto conn = std::make_shared<Communicator>(ctx);
+    int rank = ctx->Rank();
+    auto base = std::make_shared<BasicOTProtocols>(conn);
+
+    CompareProtocol comp_prot(base, radix);
+
+    [[maybe_unused]] auto b0 = ctx->GetStats()->sent_bytes.load();
+    [[maybe_unused]] auto s0 = ctx->GetStats()->sent_actions.load();
+
+    auto _c = comp_prot.Compute(inp[rank], greater_than, bw);
+
+    [[maybe_unused]] auto b1 = ctx->GetStats()->sent_bytes.load();
+    [[maybe_unused]] auto s1 = ctx->GetStats()->sent_actions.load();
+
+    SPDLOG_DEBUG(
+        "Compare {} bits {} elements sent {} bytes, {} bits each #sent {}", bw,
+        inp[0].numel(), (b1 - b0), (b1 - b0) * 8. / inp[0].numel(), (s1 - s0));
+
+    cmp_oup[rank] = _c;
+  });
+
+  DISPATCH_ALL_FIELDS(field, "", [&]() {
+    auto xout0 = NdArrayView<ring2k_t>(cmp_oup[0]);
+    auto xout1 = NdArrayView<ring2k_t>(cmp_oup[1]);
+    auto xinp0 = NdArrayView<ring2k_t>(inp[0]);
+    auto xinp1 = NdArrayView<ring2k_t>(inp[1]);
+
+    for (int64_t i = 0; i < inp[0].numel(); ++i) {
       bool expected = greater_than ? xinp0[i] > xinp1[i] : xinp0[i] < xinp1[i];
       bool got_cmp = xout0[i] ^ xout1[i];
       EXPECT_EQ(expected, got_cmp);
@@ -90,48 +156,53 @@ TEST_P(CompareProtTest, Basic) {
 
 TEST_P(CompareProtTest, WithEq) {
   size_t kWorldSize = 2;
-  int64_t n = 10;
+  Shape shape = {10, 10, 10};
   FieldType field = std::get<0>(GetParam());
   size_t radix = std::get<2>(GetParam());
   bool greater_than = std::get<1>(GetParam());
 
+  NdArrayRef _inp[2];
+  _inp[0] = ring_rand(field, shape);
+  _inp[1] = ring_rand(field, shape);
+
   NdArrayRef inp[2];
-  inp[0] = ring_rand(field, {n});
-  inp[1] = ring_rand(field, {n});
+  inp[0] = _inp[0].slice({0, 0, 0}, {10, 10, 10}, {2, 3, 2});
+  inp[1] = _inp[0].slice({0, 0, 0}, {10, 10, 10}, {2, 3, 2});
+  shape = inp[0].shape();
 
   DISPATCH_ALL_FIELDS(field, "", [&]() {
-    auto xinp = xt_mutable_adapt<ring2k_t>(inp[0]);
+    auto xinp = NdArrayView<ring2k_t>(inp[0]);
     xinp[0] = 1;
     xinp[1] = 10;
     xinp[2] = 100;
 
-    xinp = xt_mutable_adapt<ring2k_t>(inp[1]);
+    xinp = NdArrayView<ring2k_t>(inp[1]);
     xinp[0] = 1;
     xinp[1] = 9;
     xinp[2] = 1000;
   });
 
-  ArrayRef cmp_oup[2];
-  ArrayRef eq_oup[2];
+  NdArrayRef cmp_oup[2];
+  NdArrayRef eq_oup[2];
   utils::simulate(kWorldSize, [&](std::shared_ptr<yacl::link::Context> ctx) {
     auto conn = std::make_shared<Communicator>(ctx);
     int rank = ctx->Rank();
     auto base = std::make_shared<BasicOTProtocols>(conn);
     CompareProtocol comp_prot(base, radix);
-    auto [_c, _e] = comp_prot.ComputeWithEq(flatten(inp[rank]), greater_than);
+    auto [_c, _e] = comp_prot.ComputeWithEq(inp[rank], greater_than);
     cmp_oup[rank] = _c;
     eq_oup[rank] = _e;
   });
 
   DISPATCH_ALL_FIELDS(field, "", [&]() {
-    auto xout0 = ArrayView<ring2k_t>(cmp_oup[0]);
-    auto xout1 = ArrayView<ring2k_t>(cmp_oup[1]);
-    auto xeq0 = ArrayView<ring2k_t>(eq_oup[0]);
-    auto xeq1 = ArrayView<ring2k_t>(eq_oup[1]);
-    auto xinp0 = xt_adapt<ring2k_t>(inp[0]);
-    auto xinp1 = xt_adapt<ring2k_t>(inp[1]);
+    auto xout0 = NdArrayView<ring2k_t>(cmp_oup[0]);
+    auto xout1 = NdArrayView<ring2k_t>(cmp_oup[1]);
+    auto xeq0 = NdArrayView<ring2k_t>(eq_oup[0]);
+    auto xeq1 = NdArrayView<ring2k_t>(eq_oup[1]);
+    auto xinp0 = NdArrayView<ring2k_t>(inp[0]);
+    auto xinp1 = NdArrayView<ring2k_t>(inp[1]);
 
-    for (int64_t i = 0; i < n; ++i) {
+    for (int64_t i = 0; i < shape.numel(); ++i) {
       bool expected = greater_than ? xinp0[i] > xinp1[i] : xinp0[i] < xinp1[i];
       bool got_cmp = xout0[i] ^ xout1[i];
       bool got_eq = xeq0[i] ^ xeq1[i];