+++ libspu/mpc/cheetah/nonlinear/equal_prot.cc
@@ -14,15 +14,15 @@
 
 #include "libspu/mpc/cheetah/nonlinear/equal_prot.h"
 
-#include "emp-tool/utils/prg.h"
+#include "yacl/crypto/tools/prg.h"
 #include "yacl/link/link.h"
 
 #include "libspu/core/type.h"
 #include "libspu/mpc/cheetah/ot/basic_ot_prot.h"
 #include "libspu/mpc/cheetah/ot/ferret.h"
 #include "libspu/mpc/cheetah/ot/util.h"
+#include "libspu/mpc/cheetah/type.h"
 #include "libspu/mpc/common/communicator.h"
-#include "libspu/mpc/semi2k/type.h"
 #include "libspu/mpc/utils/ring_ops.h"
 
 namespace spu::mpc::cheetah {
@@ -47,25 +47,29 @@ static void SetLeafOTMsg(absl::Span<uint8_t> ot_messages, uint8_t digit,
   }
 }
 
-ArrayRef EqualProtocol::Compute(const ArrayRef& inp) {
+NdArrayRef EqualProtocol::DoCompute(const NdArrayRef& inp, size_t bit_width) {
+  SPU_ENFORCE(inp.shape().size() == 1, "need 1D array");
   auto field = inp.eltype().as<Ring2k>()->field();
-  size_t bit_width = SizeOf(field) * 8;
+  if (bit_width == 0) {
+    bit_width = SizeOf(field) * 8;
+  }
+  bit_width = std::min(bit_width, SizeOf(field) * 8);
 
-  size_t remain = bit_width % compare_radix_;
-  size_t num_digits = CeilDiv(bit_width, compare_radix_);
-  size_t radix = static_cast<size_t>(1) << compare_radix_;  // one-of-N OT
-  size_t num_cmp = inp.numel();
+  int64_t remain = bit_width % compare_radix_;
+  int64_t num_digits = CeilDiv(bit_width, compare_radix_);
+  int64_t radix = static_cast<size_t>(1) << compare_radix_;  // one-of-N OT
+  int64_t num_cmp = inp.numel();
   // init to all zero
   std::vector<uint8_t> digits(num_cmp * num_digits, 0);
 
-  DISPATCH_ALL_FIELDS(field, "", [&]() {
+  DISPATCH_ALL_FIELDS(field, "Equal_break_digits", [&]() {
     using u2k = std::make_unsigned<ring2k_t>::type;
     const auto mask_radix = makeBitsMask<u2k>(compare_radix_);
     const auto mask_remain = makeBitsMask<u2k>(remain);
-    ArrayView<u2k> xinp(inp);
+    NdArrayView<u2k> xinp(inp);
 
-    for (size_t i = 0; i < num_cmp; ++i) {
-      for (size_t j = 0; j < num_digits; ++j) {
+    for (int64_t i = 0; i < num_cmp; ++i) {
+      for (int64_t j = 0; j < num_digits; ++j) {
         uint32_t shft = j * compare_radix_;
         digits[i * num_digits + j] = (xinp[i] >> shft) & mask_radix;
         // last digits
@@ -78,8 +82,11 @@ ArrayRef EqualProtocol::Compute(const ArrayRef& inp) {
 
   std::vector<uint8_t> leaf_eq(num_cmp * num_digits, 0);
   if (is_sender_) {
-    emp::PRG prg;
-    prg.random_bool(reinterpret_cast<bool*>(leaf_eq.data()), leaf_eq.size());
+    yacl::crypto::Prg<uint8_t> prg;
+    prg.Fill(absl::MakeSpan(leaf_eq));
+    // convert u8 random to boolean random
+    std::transform(leaf_eq.begin(), leaf_eq.end(), leaf_eq.data(),
+                   [](uint8_t v) { return v & 1; });
 
     // n*M instances of 1-of-N OT
     std::vector<uint8_t> leaf_ot_msg(radix * num_cmp * num_digits, 0);
@@ -87,16 +94,16 @@ ArrayRef EqualProtocol::Compute(const ArrayRef& inp) {
     std::vector<absl::Span<uint8_t> > each_leaf_ot_msg(num_cmp * num_digits);
     for (size_t i = 0; i < each_leaf_ot_msg.size(); ++i) {
       each_leaf_ot_msg[i] =
-          absl::Span<uint8_t>{leaf_ot_msg.data() + i * radix, radix};
+          absl::MakeSpan(leaf_ot_msg.data() + i * radix, radix);
     }
 
-    for (size_t i = 0; i < num_cmp; ++i) {
+    for (int64_t i = 0; i < num_cmp; ++i) {
       auto* this_ot_msg = each_leaf_ot_msg.data() + i * num_digits;
       auto* this_digit = digits.data() + i * num_digits;
       auto* this_leaf_eq = leaf_eq.data() + i * num_digits;
 
       // Step 6, 7 of Alg1 in CF2's paper
-      for (size_t j = 0; j < num_digits; ++j) {
+      for (int64_t j = 0; j < num_digits; ++j) {
         uint8_t rnd_eq = this_leaf_eq[j] & 1;
         SetLeafOTMsg(this_ot_msg[j], this_digit[j], rnd_eq);
       }
@@ -110,10 +117,9 @@ ArrayRef EqualProtocol::Compute(const ArrayRef& inp) {
                                                absl::MakeSpan(leaf_eq), 1);
   }
 
-  using BShrTy = semi2k::BShrTy;
   auto boolean_t = makeType<BShrTy>(field, 1);
-  ArrayRef prev_eq =
-      flatten(ring_zeros(field, {static_cast<int64_t>(num_digits * num_cmp)}))
+  NdArrayRef prev_eq =
+      ring_zeros(field, {static_cast<int64_t>(num_digits * num_cmp)})
           .as(boolean_t);
 
   // Transpose from msg-major order
@@ -128,10 +134,10 @@ ArrayRef EqualProtocol::Compute(const ArrayRef& inp) {
   // m0[1], m1[1], ..., mN[1]
   // ...
   // m0[M], m1[M], ..., mN[M]
-  DISPATCH_ALL_FIELDS(field, "", [&]() {
-    ArrayView<ring2k_t> xprev_eq(prev_eq);
-    for (size_t r = 0; r < num_cmp; ++r) {
-      for (size_t c = 0; c < num_digits; ++c) {
+  DISPATCH_ALL_FIELDS(field, "Equal_transpose", [&]() {
+    NdArrayView<ring2k_t> xprev_eq(prev_eq);
+    for (int64_t r = 0; r < num_cmp; ++r) {
+      for (int64_t c = 0; c < num_digits; ++c) {
         xprev_eq[c * num_cmp + r] = leaf_eq[r * num_digits + c];
       }
     }
@@ -145,18 +151,19 @@ ArrayRef EqualProtocol::Compute(const ArrayRef& inp) {
   //     ||
   //     \/
   // eq[0], eq[1], ..., eq[N]
-  size_t current_num_digits = num_digits;
+  int64_t current_num_digits = num_digits;
   while (current_num_digits > 1) {
     // eq[i-1, j] <- eq[i, 2*j] * eq[i, 2*j+1]
-    size_t half_d = current_num_digits / 2;
-    auto lhs_eq = prev_eq.slice(0, half_d * num_cmp);
-    auto rhs_eq = prev_eq.slice(lhs_eq.numel(), lhs_eq.numel() * 2);
+    int64_t half_d = current_num_digits / 2;
+    auto lhs_eq = prev_eq.slice({0}, {half_d * num_cmp}, {1});
+    auto rhs_eq = prev_eq.slice({lhs_eq.numel()}, {lhs_eq.numel() * 2}, {1});
     SPU_ENFORCE_EQ(lhs_eq.numel(), rhs_eq.numel());
 
-    size_t remain_d = current_num_digits - 2 * half_d;
+    int64_t remain_d = current_num_digits - 2 * half_d;
     auto next_eq = basic_ot_prot_->BitwiseAnd(lhs_eq, rhs_eq);
     if (remain_d > 0) {
-      ArrayRef tmp(prev_eq.eltype(), next_eq.numel() + remain_d * num_cmp);
+      NdArrayRef tmp(prev_eq.eltype(), {next_eq.numel() + remain_d * num_cmp});
+
       std::memcpy(&tmp.at(0), &next_eq.at(0),
                   prev_eq.elsize() * next_eq.numel());
 
@@ -166,10 +173,15 @@ ArrayRef EqualProtocol::Compute(const ArrayRef& inp) {
     } else {
       prev_eq = next_eq;
     }
-    current_num_digits = CeilDiv<size_t>(prev_eq.numel(), num_cmp);
+    current_num_digits = CeilDiv(prev_eq.numel(), num_cmp);
   }
 
   return prev_eq.as(boolean_t);
+}
+
+NdArrayRef EqualProtocol::Compute(const NdArrayRef& inp, size_t bit_width) {
+  const auto& shape = inp.shape();
+  return DoCompute(inp.reshape({inp.numel()}), bit_width).reshape(shape);
 }
 
 }  // namespace spu::mpc::cheetah