+++ spu/psi_pb2.py
@@ -1,0 +1,1 @@
+from libspu.psi.psi_pb2 import *