+++ spu/spu_pb2.py
@@ -1,0 +1,1 @@
+from libspu.spu_pb2 import *