+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - [API] Add intrinsic support
 - [Feature] Support half type
 - [Feature] Add Psi Progress
+- [Feature] Add SineOp/CosineOp support
+- [Feature] Add complex support
 
 ## 20230705